// frontend/src/api/generated/models/ReportList.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ReportPublic } from './ReportPublic';
export type ReportList = {
    items: Array<ReportPublic>;
    total: number;
    offset: number;
    limit: number;
};